Modify Vendor Controller

+++ management/src/main/java/com/aditya/management/controller/VendorController.java
@@ -7,7 +7,6 @@ import com.aditya.management.constant.APiUrl;
 import com.aditya.management.entity.Vendor;
 import com.aditya.management.service.RateLimitingService;
 import com.aditya.management.service.intrface.VendorService;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -21,38 +20,43 @@ import java.util.List;
 @RequiredArgsConstructor
 @RequestMapping(path = APiUrl.VENDOR)
 public class VendorController {
-
     private final VendorService vendorService;
     private final RateLimitingService rateLimitingService;
 
     @PreAuthorize("hasAnyRole('ADMIN','USER')")
     @PostMapping
-    public ResponseEntity<BaseResponse> createVendor(@RequestBody VendorRequest request) {
+    public ResponseEntity<BaseResponse> createVendor (@RequestBody VendorRequest request) {
+        if (rateLimitingService.tryConsume()) {
+            Vendor vendor = vendorService.createVendor(request);
+            BaseResponse response = CommonResponse.<Vendor>builder()
+                    .statusCode(HttpStatus.OK.value())
+                    .message("successfully update data vendor")
+                    .data(vendor)
+                    .build();
 
-        Vendor vendor = vendorService.createVendor(request);
+            return ResponseEntity.ok(response);
 
-        BaseResponse response = CommonResponse.<Vendor>builder()
-                .statusCode(HttpStatus.OK.value())
-                .message("successfully create vendor")
-                .data(vendor)
-                .build();
-
-        return ResponseEntity.ok(response);
+        } else  {
+            throw new ResponseStatusException(HttpStatus.TOO_MANY_REQUESTS, "Rate limit exceeded, try again later!");
+        }
     }
 
     @PreAuthorize("hasAnyRole('ADMIN','USER')")
     @PutMapping
     public ResponseEntity<BaseResponse> updateVendor(@RequestBody Vendor request) {
+        if (rateLimitingService.tryConsume()) {
+            Vendor vendor =vendorService.updateVendor(request);
 
-        Vendor vendor = vendorService.updateVendor(request);
+            BaseResponse response = CommonResponse.<Vendor>builder()
+                    .statusCode(HttpStatus.OK.value())
+                    .message("successfully update data vendor")
+                    .data(vendor)
+                    .build();
 
-        BaseResponse response = CommonResponse.<Vendor>builder()
-                .statusCode(HttpStatus.OK.value())
-                .message("successfully update data vendor")
-                .data(vendor)
-                .build();
-
-        return ResponseEntity.ok(response);
+            return ResponseEntity.ok(response);
+        } else {
+            throw new ResponseStatusException(HttpStatus.TOO_MANY_REQUESTS, "Rate limit exceeded, try again later!");
+        }
     }
 
     @PreAuthorize("hasAnyRole('ADMIN','USER')")
@@ -70,41 +74,43 @@ public class VendorController {
 
             return ResponseEntity.ok(response);
         } else {
-            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body(
-                    CommonResponse.<List<Vendor>>builder()
-                            .statusCode(HttpStatus.TOO_MANY_REQUESTS.value())
-                            .message("Rate limit exceeded, try again later!")
-                            .build()
-            );}
+            throw new ResponseStatusException(HttpStatus.TOO_MANY_REQUESTS, "Rate limit exceeded, try again later!");}
     }
 
     @PreAuthorize("hasAnyRole('ADMIN','USER')")
     @GetMapping(path = APiUrl.PATH_VAR_ID)
     public ResponseEntity<BaseResponse> getVendorById (@PathVariable String id) {
+        if (rateLimitingService.tryConsume()) {
+            Vendor vendor = vendorService.getVendorById(id);
 
-        Vendor vendor = vendorService.getVendorById(id);
+            BaseResponse response = CommonResponse.<Vendor>builder()
+                    .statusCode(HttpStatus.OK.value())
+                    .message("Successfully get data vendor")
+                    .data(vendor)
+                    .build();
 
-        BaseResponse response = CommonResponse.<Vendor>builder()
-                .statusCode(HttpStatus.OK.value())
-                .message("Successfully get data vendor")
-                .data(vendor)
-                .build();
-
-        return ResponseEntity.ok(response);
+            return ResponseEntity.ok(response);
+        } else {
+            throw new ResponseStatusException(HttpStatus.TOO_MANY_REQUESTS, "Rate limit exceeded, try again later!");
+        }
     }
 
     @PreAuthorize("hasAnyRole('ADMIN','USER')")
     @DeleteMapping(path = APiUrl.PATH_VAR_ID)
     public ResponseEntity<BaseResponse> deleteVendorById (@PathVariable String id) {
+        if (rateLimitingService.tryConsume()) {
+            vendorService.deleteVendor(id);
 
-        vendorService.deleteVendor(id);
+            BaseResponse response = CommonResponse.<String>builder()
+                    .statusCode(HttpStatus.OK.value())
+                    .message("Successfully delete data vendor")
+                    .data("Deleted")
+                    .build();
 
-        BaseResponse response = CommonResponse.<String>builder()
-                .statusCode(HttpStatus.OK.value())
-                .message("Successfully delete data vendor")
-                .data("Deleted")
-                .build();
-
-        return ResponseEntity.ok(response);
+            return ResponseEntity.ok(response);
+        } else {
+            throw new ResponseStatusException(HttpStatus.TOO_MANY_REQUESTS, "Rate limit exceeded, try again later!");
+        }
     }
+
 }
